Migrar parseo de Request a trait
